1 增加播放器选择功能（待完成） 2 移除 getChannels 功能，audio subtitle 有对应方法 3 MeetPlayer 增加 MediaInfo adapter 信息 4 MediaController 移除 DisplayMode 相关接口方法（移至meetvideoview） 5 实现mediaplayer audo选择逻辑 6 timespec 修改不合适的数据类型 int64_t 7 set audio channel 失败 不会 onError

+++ engine2/player/audiorender.cpp
@@ -293,7 +293,7 @@ status_t AudioRender::render(AVFrame* audioFrame)//int16_t* buffer, uint32_t buf
 #endif
 		uint32_t sampleInCount = audioFrame->nb_samples;
 		LOGD("sampleInCount:%d", sampleInCount);
-		int sampleOutCount = av_rescale_rnd(
+		int sampleOutCount = (int)av_rescale_rnd(
 			swr_get_delay(mConvertCtx, mSampleRate) + sampleInCount,
 			mSampleRateOutput,
 			mSampleRate,
